requirements.txt
openai == 1.82.0
Flask == 3.1.1
flask-cors == 6.0.0
python-dotenv==1.0.1
